feat(user): add user existence validation in UserSaveUseCase and handle user not found in UserDeleteUseCase; update application properties

+++ src/main/java/com/x86/followup/module/user/application/usecase/UserDeleteUseCase.java
@@ -1,5 +1,6 @@
 package com.x86.followup.module.user.application.usecase;
 
+import com.x86.followup.module.user.domain.exception.UserNotFoundError;
 import com.x86.followup.module.user.domain.model.UserId;
 import com.x86.followup.module.user.domain.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -9,6 +10,13 @@ public class UserDeleteUseCase {
     private final UserRepository repository;
 
     public void execute(Integer id) {
+
+        UserId userId = new UserId(id);
+
+        if (!repository.findById(userId).isPresent()){
+            throw new UserNotFoundError("Este usuario no existe en nuestra plataforma");
+        }
+
         this.repository.delete(new UserId(id));
     }
 }
